Move `InterceptAction` to `interceptor.rs`
